+++ handy-tarifvergleich-server/Models/Form.cs
@@ -3,11 +3,11 @@
     public class Form
     {
         public bool StrongWorldWideUsage { get; set; }
-        public int CallminutesPerMonthCH { get; set; }
-        public int GBPerMonthCH { get; set; }
-        public int SMSPerMonthCH { get; set; }
-        public int CallminutesPerMonthEurope { get; set; }
-        public int GBPerMonthEurope { get; set; }
-        public int SMSPerMonthEurope { get; set; }
+        public decimal CallminutesPerMonthCH { get; set; }
+        public decimal GBPerMonthCH { get; set; }
+        public decimal SMSPerMonthCH { get; set; }
+        public decimal CallminutesPerMonthEurope { get; set; }
+        public decimal GBPerMonthEurope { get; set; }
+        public decimal SMSPerMonthEurope { get; set; }
     }
 }
