+++ handy-tarifvergleich-server/Models/OfferDeductions.cs
@@ -2,11 +2,11 @@
 {
     public class OfferDeductions
     {
-        public int FreeGBInternetCH { get; set; }
-        public int FreeCallminutesCH { get; set; }
-        public int FreeSMSCH { get; set; }
-        public int FreeGBInternetEurope { get; set; }
-        public int FreeCallminutesEurope { get; set; }
-        public int FreeSMSEurope { get; set; }
+        public decimal FreeGBInternetCH { get; set; }
+        public decimal FreeCallminutesCH { get; set; }
+        public decimal FreeSMSCH { get; set; }
+        public decimal FreeGBInternetEurope { get; set; }
+        public decimal FreeCallminutesEurope { get; set; }
+        public decimal FreeSMSEurope { get; set; }
     }
 }
